pyproject.toml
[project]
name = "agent-ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn==0.34.0",
    "fastapi==0.115.6",
    "sqlalchemy<2",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "langchain-openai==0.3.9",
    "langchain==0.3.20",
    "langchain-community==0.3.19",
    "jinja2==3.1.6",
    "loguru==0.7.3",
    "alembic==1.16.5",
    "pgvector",
    "redis==7.0.1",
    "sqlalchemy",
]

[dependency-groups]
dev = [
    "ipykernel",
    "pre-commit",
    "pytest",
]
